Renamed pull_queue.rb -> gem_map_queue.rb
