+++ gem_map_queue.rb
